+++ resources/data/login.php
@@ -10,7 +10,7 @@ $pass = $_POST['password'];
 // Check to see if our user variables are set appropriately
 if( isset($user) && isset($pass) ) {
         // Prepare our database query
-        $dbconn = $dbase->prepare('SELECT * FROM Employees WHERE Username=:username LIMIT 1;');
+        $dbconn = $dbase->prepare('SELECT * FROM Employees WHERE Username=:username AND Active = 1 LIMIT 1;');
 
         // Bind our username parameter to the query and execute it
         $dbconn->bindParam(':username', $user);